first commit by son

+++ MultiType_GC/Library/pos-malloc.h
@@ -585,7 +585,7 @@ POS_WRITE_VAUE(name, (unsigned long *)&((mchunkptr)(((char*)(p)) + (s)))->size, 
 /* Set size at footer (only when chunk is not in use) */
 #define set_foot(p, s)       (((mchunkptr)((char*)(p) + (s)))->prev_size = (s))
 //dk s
-#define set_next_seg_pointer(p, s, pp) (((mchunkptr)((char*)(p) + (s)))->size = (pp)) //size = pointer to next seg
+#define set_next_seg_pointer(p, s, pp) (((mchunkptr)((char*)(p) + (s)))->size = (unsigned long)(pp)) //size = pointer to next seg
 //dk e
 #if CONSISTENCY == 1
 #define set_head_size_log(name, p, s)   \
